+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.23.0
 
-*TBD*
+IMPROVEMENT
+- [docs] Include `ecosystem.json` from deprecated `aib-data` repository.
 
 ## 0.22.0